// clients/typescript-node/generated/model/chatCompletionRequestMessage.ts
/**
 * OpenAI API
 * APIs for sampling from and fine-tuning language models
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: blah+oapicf@cliffano.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ChatCompletionRequestMessageFunctionCall } from './chatCompletionRequestMessageFunctionCall';

export class ChatCompletionRequestMessage {
    /**
    * The role of the messages author. One of `system`, `user`, `assistant`, or `function`.
    */
    'role': ChatCompletionRequestMessage.RoleEnum;
    /**
    * The contents of the message. `content` is required for all messages except assistant messages with function calls.
    */
    'content'?: string;
    /**
    * The name of the author of this message. `name` is required if role is `function`, and it should be the name of the function whose response is in the `content`. May contain a-z, A-Z, 0-9, and underscores, with a maximum length of 64 characters.
    */
    'name'?: string;
    'functionCall'?: ChatCompletionRequestMessageFunctionCall;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "role",
            "baseName": "role",
            "type": "ChatCompletionRequestMessage.RoleEnum"
        },
        {
            "name": "content",
            "baseName": "content",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "functionCall",
            "baseName": "function_call",
            "type": "ChatCompletionRequestMessageFunctionCall"
        }    ];

    static getAttributeTypeMap() {
        return ChatCompletionRequestMessage.attributeTypeMap;
    }
}

export namespace ChatCompletionRequestMessage {
    export enum RoleEnum {
        System = <any> 'system',
        User = <any> 'user',
        Assistant = <any> 'assistant',
        Function = <any> 'function'
    }
}