// clients/typescript-node/generated/model/chatCompletionRequestMessageFunctionCall.ts
/**
 * OpenAI API
 * APIs for sampling from and fine-tuning language models
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: blah+oapicf@cliffano.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* The name and arguments of a function that should be called, as generated by the model.
*/
export class ChatCompletionRequestMessageFunctionCall {
    /**
    * The name of the function to call.
    */
    'name'?: string;
    /**
    * The arguments to call the function with, as generated by the model in JSON format. Note that the model does not always generate valid JSON, and may hallucinate parameters not defined by your function schema. Validate the arguments in your code before calling your function.
    */
    'arguments'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "arguments",
            "baseName": "arguments",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ChatCompletionRequestMessageFunctionCall.attributeTypeMap;
    }
}