// clients/typescript-node/generated/model/createChatCompletionStreamResponse.ts
/**
 * OpenAI API
 * APIs for sampling from and fine-tuning language models
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: blah+oapicf@cliffano.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CreateChatCompletionStreamResponseChoicesInner } from './createChatCompletionStreamResponseChoicesInner';

export class CreateChatCompletionStreamResponse {
    'id': string;
    'object': string;
    'created': number;
    'model': string;
    'choices': Array<CreateChatCompletionStreamResponseChoicesInner>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "object",
            "baseName": "object",
            "type": "string"
        },
        {
            "name": "created",
            "baseName": "created",
            "type": "number"
        },
        {
            "name": "model",
            "baseName": "model",
            "type": "string"
        },
        {
            "name": "choices",
            "baseName": "choices",
            "type": "Array<CreateChatCompletionStreamResponseChoicesInner>"
        }    ];

    static getAttributeTypeMap() {
        return CreateChatCompletionStreamResponse.attributeTypeMap;
    }
}